I've dug myself a hole with functionality...
Right now, I am at an impasse where I must decide what I do further.

The open function takes regex or it takes a string. Either I must build
functionality that it's intended to match, or that it's intended to open
a new file, or both.

I want to do both, but then that means I have to do input parsing to
make sure I am not receiving a regex string, and that's not fair...
perhaps the flag should be "match-many" or something