Removing helpers, migrated those functions into Page
Some caveats: could not get page_op to work as a static method of a class. I supposed there are unwritten rules for writing static classes with blocks that I am unaware of. For now, I will leave page_op as a funciton until I figure how I want to expand it... it may mean a refactor, but that is something I am okay with